requirements.txt
antlr4-python3-runtime==4.9.3
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
bitmex==0.2.2
bravado==11.0.3
bravado-core==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
config==0.5.1
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
Django==4.2.7
django-environ==0.11.2
executing==2.0.1
fonttools==4.44.0
fqdn==1.5.1
gunicorn==21.2.0
hydra-core==1.3.2
idna==3.4
ipython==8.17.2
isoduration==20.11.0
jedi==0.19.1
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
llvmlite==0.41.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
monotonic==1.6
msgpack==1.0.7
numba==0.58.1
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
update==0.0.1
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.9
webcolors==1.13

python-dotenv~=1.0.0
DateTime~=5.1
numpy~=1.23.5
pandas~=2.0.1
python-dateutil~=2.8.2
tqdm~=4.65.0
omegaconf~=2.3.0
PyYAML~=6.0.1